chore: cleanup and simplify
